Use config to control direct school DfE Sign-in
