+++ .env.example
@@ -10,6 +10,7 @@ ENABLE_PERSONAS=true
 ENABLE_MIGRATION_TESTING=false
 ENABLE_MIGRATION_INTERFACE=false
 ENABLE_SCHOOLS_INTERFACE=false
+ENABLE_DIRECT_SCHOOL_DFE_SIGN_IN=false
 ENABLE_OTP_SCHOOL_SIGN_IN=false
 ENABLE_HUBS=false
 ENABLE_INCIDENT_BANNER=false
