+++ app/views/appropriate_bodies/landing/show.md.erb
@@ -10,7 +10,7 @@ You’ll need to give details about:
 * the induction history and progress of your ECTs
 * whether an ECT has passed or failed, or had their induction extended
 
-<% if Rails.application.config.dfe_sign_in_enabled && !Rails.application.config.enable_personas %>
+<% if direct_dfe_sign_in_available? %>
   <%= govuk_start_button(text: "Start now", href: '/auth/dfe_sign_in', as_button: true) %>
 <% else %>
   <%= govuk_start_button(text: "Start now", href: sign_in_path) %>
