+++ config/application.rb
@@ -53,6 +53,7 @@ module RegisterEarlyCareerTeachers
     config.enable_migration_interface = ActiveModel::Type::Boolean.new.cast(ENV.fetch("ENABLE_MIGRATION_INTERFACE", false))
     config.enable_migration_testing = ActiveModel::Type::Boolean.new.cast(ENV.fetch("ENABLE_MIGRATION_TESTING", false))
     config.enable_schools_interface = ActiveModel::Type::Boolean.new.cast(ENV.fetch("ENABLE_SCHOOLS_INTERFACE", false))
+    config.enable_direct_school_dfe_sign_in = ActiveModel::Type::Boolean.new.cast(ENV.fetch("ENABLE_DIRECT_SCHOOL_DFE_SIGN_IN", false))
     config.enable_otp_school_sign_in = ActiveModel::Type::Boolean.new.cast(ENV.fetch("ENABLE_OTP_SCHOOL_SIGN_IN", false))
     config.enable_teaching_school_hubs = ActiveModel::Type::Boolean.new.cast(ENV.fetch("ENABLE_HUBS", false))
     config.enable_incident_banner = ActiveModel::Type::Boolean.new.cast(ENV.fetch("ENABLE_INCIDENT_BANNER", false))
