+++ config/terraform/application/config/production.yml
@@ -4,6 +4,7 @@ ENABLE_SENTRY: true
 SENTRY_ENVIRONMENT: production
 ENABLE_PERSONAS: false
 ENABLE_FAKE_TRS_API: false
+ENABLE_DIRECT_SCHOOL_DFE_SIGN_IN: true
 ENABLE_BLAZER: true
 ENABLE_API: false
 ENABLE_MIGRATION_TESTING: false
